backend/requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
cached-property==1.5.2
celery==5.1.2
click==8.0.1
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.6
django-cors-headers==3.8.0
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
gunicorn==20.1.0
importlib-metadata==4.7.1
kombu==5.1.0
mysqlclient==2.0.3
prompt-toolkit==3.0.20
psycopg2-binary==2.9.1
PyJWT==1.7.1
pytz==2021.1
redis==3.5.3
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
vine==5.0.0
wcwidth==0.2.5
zipp==3.5.0
